Component rearrange
- Extract SessionInfoBar to separate component
- App for Redux, Router, and main Layout
- src/index.tsx for main entrypoint
- remove index.js at top level of project
