--- index.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { render } from "react-dom";
-
-import KintoAdmin from "./src";
-
-render(<KintoAdmin />, document.getElementById("app"));
